requirements.txt
# SaFeR-Steer Dependencies

# Core
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.28.0
datasets>=2.18.0

# Vision
Pillow>=10.0.0
torchvision>=0.15.0

# API & Serving
openai>=1.12.0
vllm>=0.4.0
requests>=2.31.0

# Training
deepspeed>=0.14.0
peft>=0.10.0
trl>=0.8.0

# Evaluation
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0

# Utils
pyyaml>=6.0
jsonlines>=4.0.0
